+++ tanjun/clients.py
@@ -810,9 +810,9 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
 
     async def __aexit__(
         self,
-        exception_type: typing.Optional[type[BaseException]],
-        exception: typing.Optional[BaseException],
-        exception_traceback: typing.Optional[types.TracebackType],
+        exc_type: typing.Optional[type[Exception]],
+        exc: typing.Optional[Exception],
+        exc_traceback: typing.Optional[types.TracebackType],
     ) -> None:
         await self.close()
 
@@ -1866,7 +1866,10 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
 
             _LOGGER.info("Loading from %s", module_path)
             module = yield lambda: importlib.import_module(module_path)
-            self._call_loaders(module_path, _get_loaders(module, module_path))
+
+            with _WrapLoadError(errors.FailedModuleLoad):
+                self._call_loaders(module_path, _get_loaders(module, module_path))
+
             self._modules[module_path] = module
 
         else:
@@ -1876,7 +1879,10 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
 
             _LOGGER.info("Loading from %s", module_path)
             module = yield lambda: _get_path_module(module_path)
-            self._call_loaders(module_path, _get_loaders(module, module_path))
+
+            with _WrapLoadError(errors.FailedModuleLoad):
+                self._call_loaders(module_path, _get_loaders(module, module_path))
+
             self._path_modules[module_path_abs] = module
 
     def load_modules(self: _ClientT, *modules: typing.Union[str, pathlib.Path]) -> _ClientT:
@@ -1886,13 +1892,14 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
                 module_path = module_path.absolute()
 
             generator = self._load_module(module_path)
-            module = next(generator)()
+            load_module = next(generator)
+            with _WrapLoadError(errors.FailedModuleLoad):
+                module = load_module()
+
             try:
                 generator.send(module)
-
             except StopIteration:
                 pass
-
             else:
                 raise RuntimeError("Generator didn't finish")
 
@@ -1906,13 +1913,14 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
                 module_path = await loop.run_in_executor(None, module_path.absolute)
 
             generator = self._load_module(module_path)
-            module = await loop.run_in_executor(None, next(generator))
+            load_module = next(generator)
+            with _WrapLoadError(errors.FailedModuleLoad):
+                module = await loop.run_in_executor(None, load_module)
+
             try:
                 generator.send(module)
-
             except StopIteration:
                 pass
-
             else:
                 raise RuntimeError("Generator didn't finish")
 
@@ -1920,17 +1928,22 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
         # <<inherited docstring from tanjun.ab.Client>>.
         for module_path in modules:
             if isinstance(module_path, str):
-                module = self._modules.pop(module_path, None)
+                modules_dict: dict[typing.Any, types.ModuleType] = self._modules
 
             else:
+                modules_dict = self._path_modules
                 module_path = module_path.absolute()
-                module = self._path_modules.pop(module_path, None)
 
+            module = modules_dict.get(module_path)
             if not module:
                 raise errors.ModuleStateConflict(f"Module {module_path!s} not loaded", module_path)
 
             _LOGGER.info("Unloading from %s", module_path)
-            self._call_unloaders(module_path, _get_loaders(module, module_path))
+            with _WrapLoadError(errors.FailedModuleUnload):
+                self._call_unloaders(module_path, _get_loaders(module, module_path))
+
+            # TODO: test this isn't hit when there's no unloaders
+            del modules_dict[module_path]
 
         return self
 
@@ -1960,25 +1973,31 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
 
         _LOGGER.info("Reloading %s", module_path)
 
+        old_loaders = _get_loaders(old_module, module_path)
+        # We assert that the old module has unloaders early to avoid unnecessarily
+        # importing the new module.
+        if not any(loader.has_unload for loader in old_loaders):
+            raise errors.ModuleMissingLoaders(f"Didn't find any unloaders in old {module_path}", module_path)
+
         module = yield load_module
 
         loaders = _get_loaders(module, module_path)
-        old_loaders = _get_loaders(old_module, module_path)
 
         # We assert that the new module has loaders early to avoid unnecessarily
         # unloading then rolling back when we know it's going to fail to load.
         if not any(loader.has_load for loader in loaders):
             raise errors.ModuleMissingLoaders(f"Didn't find any loaders in new {module_path}", module_path)
 
-        self._call_unloaders(module_path, old_loaders)
+        with _WrapLoadError(errors.FailedModuleUnload):
+            # This will never raise MissingLoaders as we assert this earlier
+            self._call_unloaders(module_path, old_loaders)
 
         try:
+            # This will never raise MissingLoaders as we assert this earlier
             self._call_loaders(module_path, loaders)
-
-        except Exception:
+        except Exception as exc:
             self._call_loaders(module_path, old_loaders)
-            raise
-
+            raise errors.FailedModuleLoad from exc
         else:
             modules_dict[module_path] = module
 
@@ -1989,13 +2008,14 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
                 module_path = module_path.absolute()
 
             generator = self._reload_module(module_path)
-            module = next(generator)()
+            load_module = next(generator)
+            with _WrapLoadError(errors.FailedModuleLoad):
+                module = load_module()
+
             try:
                 generator.send(module)
-
             except StopIteration:
                 pass
-
             else:
                 raise RuntimeError("Generator didn't finish")
 
@@ -2009,7 +2029,10 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
                 module_path = await loop.run_in_executor(None, module_path.absolute)
 
             generator = self._reload_module(module_path)
-            module = await loop.run_in_executor(None, next(generator))
+            load_module = next(generator)
+            with _WrapLoadError(errors.FailedModuleLoad):
+                module = await loop.run_in_executor(None, load_module)
+
             try:
                 generator.send(module)
 
@@ -2207,3 +2230,22 @@ def _get_path_module(module_path: pathlib.Path, /) -> types.ModuleType:
     module = importlib_util.module_from_spec(spec)
     spec.loader.exec_module(module)
     return module
+
+
+class _WrapLoadError:
+    __slots__ = ("_error",)
+
+    def __init__(self, error: collections.Callable[[], Exception], /) -> None:
+        self._error = error
+
+    def __enter__(self) -> None:
+        pass
+
+    def __exit__(
+        self,
+        exc_type: typing.Optional[type[Exception]],
+        exc: typing.Optional[Exception],
+        exc_tb: typing.Optional[types.TracebackType],
+    ) -> None:
+        if exc and not isinstance(exc, errors.ModuleMissingLoaders):
+            raise self._error() from exc  # noqa: R102 unnecessary parenthesis on raised exception